feat: Updated UserForm test file
Added import statements and uncommented test cases
